[#36] refactor: 패키지 구조 리팩토링
DTO 패키지 domain => application
