+++ src/main/java/dabang/star/cafe/api/CurrentMemberApi.java
@@ -5,10 +5,10 @@ import dabang.star.cafe.api.aop.LoginCheck;
 import dabang.star.cafe.api.aop.SessionId;
 import dabang.star.cafe.api.request.CurrentMemberRequest;
 import dabang.star.cafe.api.request.MemberUpdateRequest;
-import dabang.star.cafe.domain.member.MemberData;
+import dabang.star.cafe.application.MemberService;
+import dabang.star.cafe.application.data.MemberData;
 import dabang.star.cafe.domain.authentication.MemberLoginService;
 import dabang.star.cafe.domain.member.Member;
-import dabang.star.cafe.application.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
