+++ src/main/java/dabang/star/cafe/domain/authentication/MemberLoginService.java
@@ -1,7 +1,7 @@
 package dabang.star.cafe.domain.authentication;
 
 
-import dabang.star.cafe.domain.member.MemberData;
+import dabang.star.cafe.application.data.MemberData;
 
 public interface MemberLoginService {
