+++ src/main/java/dabang/star/cafe/infrastructure/repository/MybatisManagerRepository.java
@@ -1,7 +1,7 @@
 package dabang.star.cafe.infrastructure.repository;
 
 import dabang.star.cafe.domain.manager.Manager;
-import dabang.star.cafe.domain.manager.ManagerData;
+import dabang.star.cafe.application.data.ManagerData;
 import dabang.star.cafe.domain.manager.ManagerRepository;
 import dabang.star.cafe.infrastructure.mapper.ManagerMapper;
 import dabang.star.cafe.utils.page.Page;
