+++ src/main/java/dabang/star/cafe/infrastructure/repository/MybatisMemberRepository.java
@@ -1,6 +1,6 @@
 package dabang.star.cafe.infrastructure.repository;
 
-import dabang.star.cafe.domain.member.MemberData;
+import dabang.star.cafe.application.data.MemberData;
 import dabang.star.cafe.domain.member.Member;
 import dabang.star.cafe.domain.member.MemberRepository;
 import dabang.star.cafe.infrastructure.mapper.MemberMapper;
