+++ src/main/java/dabang/star/cafe/infrastructure/service/SessionManagerLoginService.java
@@ -3,7 +3,7 @@ package dabang.star.cafe.infrastructure.service;
 import dabang.star.cafe.api.exception.NoAuthenticationException;
 import dabang.star.cafe.domain.authentication.EncryptService;
 import dabang.star.cafe.domain.authentication.ManagerLoginService;
-import dabang.star.cafe.domain.manager.ManagerData;
+import dabang.star.cafe.application.data.ManagerData;
 import dabang.star.cafe.domain.manager.ManagerRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
