+++ src/main/java/dabang/star/cafe/infrastructure/service/SessionMemberLoginService.java
@@ -3,7 +3,7 @@ package dabang.star.cafe.infrastructure.service;
 import dabang.star.cafe.api.exception.NoAuthenticationException;
 import dabang.star.cafe.domain.authentication.EncryptService;
 import dabang.star.cafe.domain.authentication.MemberLoginService;
-import dabang.star.cafe.domain.member.MemberData;
+import dabang.star.cafe.application.data.MemberData;
 import dabang.star.cafe.domain.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
